FeatureTransforms

FeatureTransforms.AbstractScaling

FeatureTransforms.HoD

FeatureTransforms.IdentityScaling

FeatureTransforms.LinearCombination

FeatureTransforms.MeanStdScaling

FeatureTransforms.OneHotEncoding

FeatureTransforms.Periodic

FeatureTransforms.Periodic

FeatureTransforms.Power

FeatureTransforms.Transform

FeatureTransforms._periodic

FeatureTransforms._try_copy

FeatureTransforms.apply

FeatureTransforms.apply

FeatureTransforms.apply

FeatureTransforms.apply

FeatureTransforms.apply

FeatureTransforms.apply

FeatureTransforms.apply!

FeatureTransforms.apply!

FeatureTransforms.apply!

FeatureTransforms.transform

FeatureTransforms.transform!

FeatureTransforms.AbstractScaling — Type

AbstractScaling <: Transform

Linearly scale the data according to some statistics.

FeatureTransforms.HoD — Type

HoD <: Transform

Get the hour of day corresponding to the data.

FeatureTransforms.IdentityScaling — Type

IdentityScaling <: AbstractScaling

Represents the no-op scaling which simply returns the data it is applied on.

FeatureTransforms.LinearCombination — Type

LinearCombination(coefficients) <: Transform

Calculate the linear combination using the vector coefficients passed in.

FeatureTransforms.MeanStdScaling — Type

MeanStdScaling(μ, σ) <: AbstractScaling

Linearly scale the data by the statistical mean μ and standard deviation σ. This is also known as standardization, or the Z score transform.

Keyword arguments to apply

inverse=true: inverts the scaling (e.g. to reconstruct the unscaled data).

eps=1e-3: used in place of all 0 values in σ before scaling (if inverse=false).

FeatureTransforms.OneHotEncoding — Type

OneHotEncoding{R<:Real} <: Transform

One-hot encode the categorical value for each target element.

Construct a n-by-p binary matrix, given a Vector of target data x (of length n) and a Vector of all unique possible values in x (of length p).

The element [i, j] is true if the i^th target in x corresponds to the j^th possible value and false otherwise. Note that Rcan be specified to determine the return type of results. It defaults to a Matrix of Bools.

Note that this Transform does not support specifying dims other than : (all dims) because it is a one-to-many transform (for example a Vector input produces a Matrix output).

FeatureTransforms.Periodic — Type

Periodic{P, S}(f, period::P, [phase_shift::S]) <: Transform

Applies a periodic function f with provided period and phase_shift to the data.

The period and phase_shift must have the same supertype of Real or Period, depending on whether the data is Real or TimeType respectively.

Note

For TimeType data, the result will change depending on the type of period given, even if the same amount of time is described. Example: Week(1) vs Second(Week(1)); the former starts the period on the most recent Monday, while the latter starts the period on the most recent multiple of 604800 seconds since time 0.

Fields

f::Union{typeof(cos), typeof(sin)}: the periodic function

period::Union{Real, Period}: the function period. Must be strictly positive.

phase_shift::Union{Real, Period} (optional): adjusts the phase of the periodic function, measured in the same units as the input. Increasing the value translates the function to the right, toward higher/later input values.

FeatureTransforms.Periodic — Method

Periodic(f, period) -> Periodic

A constructor for Periodic. Returns a Periodic transform with zero phase shift.

FeatureTransforms.Power — Type

Power(exponent) <: Transform

Raise the data by the given exponent.

FeatureTransforms.Transform — Type

Transform

Abstract supertype for all feature Transforms.

FeatureTransforms._periodic — Function

_periodic(f, instant, period, phase_shift=Day(0))

Computes the value of periodic function f at the given instant in time.

Arguments

f: the periodic function

period: the function period

phase_shift: adjusts the phase of the periodic function. Increasing the value translates the function to the right, toward higher/later input values.

FeatureTransforms._try_copy — Method

_try_copy(data)

Try to copy the data, fallback to deepcopy if not supported. Not all objects support copy, but we should use it to improve performance if possible.

FeatureTransforms.apply — Function

apply(data::T, ::Transform; kwargs...)

Applies the Transform to the data. New transforms should usually only extend _apply which this method delegates to.

Where necessary, this should be extended for new data types T.

FeatureTransforms.apply! — Function

apply!(data::T, ::Transform; kwargs...) -> T

Applies the Transform mutating the input data. This method delegates to apply under the hood so does not need to be defined separately.

If Transform does not support mutation, this method will error.

FeatureTransforms.apply! — Method

apply!(A::AbstractArray, ::Transform; dims=:, kwargs...)

Applies the Transform to each element of A, mutating the data.

FeatureTransforms.apply! — Method

apply!(table::T, ::Transform; cols=nothing)::T where T

Applies the Transform to each of the specified columns in the table. If no cols are specified, then the Transform is applied to all columns.

FeatureTransforms.apply — Method

apply(A::AbstractArray, ::Transform; dims=:, inds=:, kwargs...)

Applies the Transform to the elements of A.

Provide the dims keyword to apply the Transform along a certain dimension. For example, given a Matrix, dims=1 applies to each column, while dims=2 applies to each row.

Provide the inds keyword to apply the Transform to certain indices along the dims specified.

Note: if dims === : (all dimensions), then inds will be the global indices of the array, instead of being relative to a certain dimension.

This method does not guarantee the data type of what is returned. It will try to conserve type but the returned type depends on what the original A was, and the dims and inds specified.

FeatureTransforms.apply — Method

apply(x::AbstractVector, LC::LinearCombination; inds=Colon())

Applies the LinearCombination to each of the specified indices in x.

If no inds are specified, then the LinearCombination is applied to all elements.

FeatureTransforms.apply — Method

apply(x::AbstractMatrix, LC::LinearCombination; dims=1, inds=Colon())

Applies the LinearCombination to each of the specified indices in x along the dimension specified, which defaults to applying it row-wise for each column of x.

If no inds are specified, then the LinearCombination is applied to all columns.

FeatureTransforms.apply — Method

apply(x::Table, LC::LinearCombination; cols=nothing)

Applies the LinearCombination to each of the specified cols in x.

If no cols are specified, then the LinearCombination is applied to all columns.

FeatureTransforms.apply — Method

apply(table, ::Transform; cols=nothing, kwargs...) -> Vector

Applies the Transform to each of the specified columns in the table. If no cols are specified, then the Transform is applied to all columns.

Return

If cols is a single value (not in a list): the transformed column vector.

Otherwise: an array containing each transformed column, in the same order as cols.

FeatureTransforms.transform — Function

transform(::T, data)

Non-mutating version of transform!.

FeatureTransforms.transform! — Function

transform!(::T, data)

Defines the feature engineering pipeline for some type T, which comprises a collection of Transforms to be peformed on the data.

transform! should be overloaded for custom types T that require feature engineering.