API

Transforms

FeatureTransforms.Transform — Type

Transform

Abstract supertype for all feature Transforms.

FeatureTransforms.HoD — Type

HoD <: Transform

Get the hour of day corresponding to the data.

FeatureTransforms.Power — Type

Power(exponent) <: Transform

Raise the data by the given exponent.

FeatureTransforms.Periodic — Type

Periodic{P, S}(f, period::P, [phase_shift::S]) <: Transform

Applies a periodic function f with provided period and phase_shift to the data.

The period and phase_shift must have the same supertype of Real or Period, depending on whether the data is Real or TimeType respectively.

Note

For TimeType data, the result will change depending on the type of period given, even if the same amount of time is described. Example: Week(1) vs Second(Week(1)); the former starts the period on the most recent Monday, while the latter starts the period on the most recent multiple of 604800 seconds since time 0.

Fields

f::Union{typeof(cos), typeof(sin)}: the periodic function

period::Union{Real, Period}: the function period. Must be strictly positive.

phase_shift::Union{Real, Period} (optional): adjusts the phase of the periodic function, measured in the same units as the input. Increasing the value translates the function to the right, toward higher/later input values.

FeatureTransforms.AbstractScaling — Type

AbstractScaling <: Transform

Linearly scale the data according to some statistics.

FeatureTransforms.MeanStdScaling — Type

MeanStdScaling(μ, σ) <: AbstractScaling

Linearly scale the data by the statistical mean μ and standard deviation σ. This is also known as standardization, or the Z score transform.

Keyword arguments to apply

inverse=true: inverts the scaling (e.g. to reconstruct the unscaled data).

eps=1e-3: used in place of all 0 values in σ before scaling (if inverse=false).

FeatureTransforms.IdentityScaling — Type

IdentityScaling <: AbstractScaling

Represents the no-op scaling which simply returns the data it is applied on.

FeatureTransforms.LinearCombination — Type

LinearCombination(coefficients) <: Transform

Calculate the linear combination using the vector coefficients passed in.

FeatureTransforms.OneHotEncoding — Type

OneHotEncoding{R<:Real} <: Transform

One-hot encode the categorical value for each target element.

Construct a n-by-p binary matrix, given a Vector of target data x (of length n) and a Vector of all unique possible values in x (of length p).

The element [i, j] is true if the i^th target in x corresponds to the j^th possible value and false otherwise. Note that Rcan be specified to determine the return type of results. It defaults to a Matrix of Bools.

Note that this Transform does not support specifying dims other than : (all dims) because it is a one-to-many transform (for example a Vector input produces a Matrix output).

Applying Transforms

FeatureTransforms.apply — Function

apply(data::T, ::Transform; kwargs...)

Applies the Transform to the data. New transforms should usually only extend _apply which this method delegates to.

Where necessary, this should be extended for new data types T.

FeatureTransforms.apply! — Function

apply!(data::T, ::Transform; kwargs...) -> T

Applies the Transform mutating the input data. This method delegates to apply under the hood so does not need to be defined separately.

If Transform does not support mutation, this method will error.

FeatureTransforms.transform! — Function

transform!(::T, data)

Defines the feature engineering pipeline for some type T, which comprises a collection of Transforms to be peformed on the data.

transform! should be overloaded for custom types T that require feature engineering.

FeatureTransforms.transform — Function

transform(::T, data)

Non-mutating version of transform!.