TestUtils

FeatureTransforms.TestUtils is used to test new data types that wish to support the transform interface described in the documentation. It provides various test fakes and utilities to help with doing so.

API

FeatureTransforms.TestUtils.FakeManyToManyTransform — Type

FeatureTransforms.TestUtils.FakeManyToManyTransform <: Transform

A fake FeatureTransforms.ManyToMany transform for test purposes. Calling apply will return an array of ones with a size and dimension matching the cardinality of the transform.

FeatureTransforms.TestUtils.FakeManyToOneTransform — Type

FeatureTransforms.TestUtils.FakeManyToOneTransform <: Transform

A fake FeatureTransforms.ManyToOne transform for test purposes. Calling apply will return an array of ones with a size and dimension matching the cardinality of the transform.

FeatureTransforms.TestUtils.FakeOneToManyTransform — Type

FeatureTransforms.TestUtils.FakeOneToManyTransform <: Transform

A fake FeatureTransforms.OneToMany transform for test purposes. Calling apply will return an array of ones with a size and dimension matching the cardinality of the transform.

FeatureTransforms.TestUtils.FakeOneToOneTransform — Type

FeatureTransforms.TestUtils.FakeOneToOneTransform <: Transform

A fake FeatureTransforms.OneToOne transform for test purposes. Calling apply will return an array of ones with a size and dimension matching the cardinality of the transform.

FeatureTransforms.TestUtils.is_transformable — Method

is_transformable(x)

Determine if x is both a valid input and output of any Transform, i.e. that it has an apply method defined and therefore follows the transform interface.