Transforms

Transforms.HoD

Transforms.LinearCombination

Transforms.MeanStdScaling

Transforms.MeanStdScaling

Transforms.OneHotEncoding

Transforms.Periodic

Transforms.Periodic

Transforms.Power

Transforms.Scaling

Transforms.Transform

Transforms._periodic

Transforms._try_copy

Transforms.apply

Transforms.apply

Transforms.apply

Transforms.apply

Transforms.apply

Transforms.apply

Transforms.apply!

Transforms.apply!

Transforms.apply!

Transforms.transform

Transforms.transform!

Transforms.HoD — Type

HoD <: Transform

Get the hour of day corresponding to the data.

Transforms.LinearCombination — Type

LinearCombination(coefficients) <: Transform

Calculate the linear combination using the vector coefficients passed in.

Transforms.MeanStdScaling — Type

MeanStdScaling(mean, std) <: Scaling

Linearly scale the data by a statistical mean and standard deviation std. This is also known as standardization, or the Z score transform. Once computed, the statistics of a MeanStdScaling are immutable.

Can take a precomputed mean and std as arguments, or compute them from data.

Arguments

mean::NamedTuple: tuple of mean values, named by the scope of values it applies to. (all=μ, ) will apply to all data; (1=μ1, 2=μ2) for AbstractArray data will apply μ1 to the first slice and μ2 to the second slice; (a=μ1, b=μ2) for Table data will apply μ1 to column a and μ2 to column b.

std::NamedTuple: similar to mean but for standard deviation values.

Keyword arguments to apply

inverse=true: inverts the scaling (e.g. to reconstruct the unscaled data)

eps=1e-3: replaces all 0 values in std before scaling (if inverse=false)

Transforms.MeanStdScaling — Method

MeanStdScaling(data; kwargs...) <: Scaling

Construct a MeanStdScaling using the mean and standard deviation of the data.

Note

dims and cols keyword arguments must be specified the same way when constructing and applying the transform. Otherwise, the results will be inconsistent, or an error may occur.

Keyword arguments

dims=:: for AbstractArray data, the dimension(s) to compute statistics along.

cols=nothing: for Table data, the column names to compute statistics for.

Transforms.OneHotEncoding — Type

OneHotEncoding <: Transform

One-hot encode the categorical value for each target element.

Construct a n-by-p binary matrix, given a Vector of target data x (of length n) and a Vector of all unique possible values in x (of length p).

The element [i, j] is 1 if the i^th target in x corresponds to the j^th possible value and 0 otherwise.

Note that this Transform does not support specifying dims other than : (all dims) because it is a one-to-many transform (for example a Vector input produces a Matrix output).

Transforms.Periodic — Type

Periodic{T}(f, period::T, phase_shift::T) <: Transform

Applies a periodic function f with provided period and phase_shift to the data.

Note

For TimeType data, the result will change depending on the type of period given, even if the same amount of time is described. Example: Week(1) vs Second(Week(1)); the former starts the period on the most recent Monday, while the latter starts the period on the most recent multiple of 604800 seconds since time 0.

Fields

f::Union{typeof(cos), typeof(sin)}: the periodic function

period::Union{Real, Period}: the function period. Must be strictly positive.

phase_shift::Union{Real, Period}: adjusts the phase of the periodic function, measured in the same units as the input. Increasing the value translates the function to the right, toward higher/later input values.

Transforms.Periodic — Method

Periodic(f, period) -> Periodic

A constructor for Periodic. Returns a Periodic transform with zero phase shift.

Transforms.Power — Type

Power(exponent) <: Transform

Raise the data by the given exponent.

Transforms.Scaling — Type

Scaling <: Transform

Linearly scale the data as ax + b, according to some statistics a and b.

Transforms.Transform — Type

Transform

Abstract supertype for all Transforms.

Transforms._periodic — Function

_periodic(f, instant, period, phase_shift=Day(0))

Computes the value of periodic function f at the given instant in time.

Arguments

f: the periodic function

period: the function period

phase_shift: adjusts the phase of the periodic function. Increasing the value translates the function to the right, toward higher/later input values.

Transforms._try_copy — Method

_try_copy(data)

Try to copy the data, fallback to deepcopy if not supported. Not all objects support copy, but we should use it to improve performance if possible.

Transforms.apply — Function

apply(data::T, ::Transform; kwargs...)

Applies the Transform to the data. New transforms should usually only extend _apply which this method delegates to.

Where necessary, this should be extended for new data types T.

Transforms.apply! — Function

apply!(data::T, ::Transform; kwargs...) -> T

Applies the Transform mutating the input data. New transforms should usually only extend _apply! which this method delegates to.

Where necessary, this should be extended for new data types T.

Transforms.apply! — Method

apply!(A::AbstractArray, ::Transform; dims=:, kwargs...)

Applies the Transform to each element of A. Optionally specify the dims to apply the Transform along certain dimensions. For example in a Matrix, dims=1 applies to each column, while dims=2 applies to each row.

Note

For arrays with more than 2 dimensions, single dims are not supported.

Transforms.apply! — Method

apply!(table::T, ::Transform; cols=nothing)::T where T

Applies the Transform to each of the specified columns in the table. If no cols are specified, then the Transform is applied to all columns.

Transforms.apply — Method

apply(A::AbstractArray, ::Transform; dims=:, inds=:, kwargs...)

Applies the Transform to the elements of A. Provide the dims keyword to apply the Transform along a certain dimension. Provide the inds keyword to apply the Transform to certain indices along the dims specified.

Note: if dims === : (all dimensions), then inds will be the global indices of the array, instead of being relative to a certain dimension.

This method does not guarantee the data type of what is returned. It will try to conserve type but the returned type depends on what the original A was, and the dims and inds specified.

Transforms.apply — Method

apply(x::AbstractVector, LC::LinearCombination; inds=Colon())

Applies the LinearCombination to each of the specified indices in x.

If no inds are specified, then the LinearCombination is applied to all elements.

Transforms.apply — Method

apply(x::AbstractMatrix, LC::LinearCombination; dims=1, inds=Colon())

Applies the LinearCombination to each of the specified indices in x along the dimension specified, which defaults to applying it row-wise for each column of x.

If no inds are specified, then the LinearCombination is applied to all columns.

Transforms.apply — Method

apply(x::Table, LC::LinearCombination; cols=nothing)

Applies the LinearCombination to each of the specified cols in x.

If no cols are specified, then the LinearCombination is applied to all columns.

Transforms.apply — Method

apply(table, ::Transform; cols=nothing, kwargs...) -> Vector

Applies the Transform to each of the specified columns in the table. If no cols are specified, then the Transform is applied to all columns.

Returns an array containing each transformed column, in the same order as cols.

Transforms.transform — Function

transform(::T, data)

Non-mutating version of transform!.

Transforms.transform! — Function

transform!(::T, data)

Defines the feature engineering pipeline for some type T, which comprises a collection of Transforms to be peformed on the data.

transform! should be overloaded for custom types T that require feature engineering.